A refused undo says so, and the completion tests get their own file
Two corrections from review.

THE UNDO FAILED SILENTLY, and the row leaving is what hid it. The completion's
refetch drops the row, the component unmounts, and React Query drops the
observer that a per-call onError hangs off — so a refused undo showed nothing at
all. The reader pressed the one control that could take back their misclick, it
was refused, and the screen said nothing. It reports from a promise this closure
owns now (mutateAsync), which survives the unmount.

The test that should have caught it could not: the testkit's stub answers the
same static queue however many writes land, so the row never actually left and
the unmounted state was never reached. The new test drives a stub that empties
the queue after the completion, which is the state the defect lives in.

worklist.test.tsx had reached 1067 lines against the tree's 1000 ceiling. The
completion tests move to their own file, and the seam is the subject rather than
the line count: everything there is about ONE verb and what follows it, while
the file it left covers what the queue draws.

Mutation-checked: restoring the per-call onError fails the new test.
